requirements.txt
pytest==7.2
selene>=2.0.0b4
selenium~=4.8.0
allure-pytest~=2.12.0
allure-python-commons~=2.12.0
python-dotenv~=0.21.0